+++ cuda_core/tests/test_memory.py
@@ -38,7 +38,7 @@ from cuda.core._dlpack import DLDeviceType
 from cuda.core._memory import IPCBufferDescriptor
 from cuda.core._utils.cuda_utils import CUDAError, handle_return
 from cuda.core.utils import StridedMemoryView
-from helpers import IS_WINDOWS
+from helpers import IS_WINDOWS, supports_ipc_mempool
 from helpers.buffers import DummyUnifiedMemoryResource
 
 from conftest import (
@@ -46,7 +46,6 @@ from conftest import (
     skip_if_managed_memory_unsupported,
     skip_if_pinned_memory_unsupported,
 )
-from cuda_python_test_helpers import supports_ipc_mempool
 
 POOL_SIZE = 2097152  # 2MB size
